+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/CryptoPrivateKey.kt
@@ -214,8 +214,9 @@ sealed class CryptoPrivateKey<T : CryptoPublicKey>(
         val encodeCurve: Boolean,
         publicKey: CryptoPublicKey.EC?,
         attributes: List<Asn1Element>? = null
-    ) :
-        CryptoPrivateKey<CryptoPublicKey.EC>(attributes) {
+    ) :        CryptoPrivateKey<CryptoPublicKey.EC>(attributes) {
+
+
 
         override val oid = EC.oid