+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 * Add **Private Key Representation** to `indispensable`
     * Parsing of PEM and DER-encoded private keys
     * Introduce `SignatureAlgorithm.signerFor(privateKey)` to create signers backed by (previously parsed, or manually constructed) private keys
+    * Export of private keys from ephemeral signers (and only ephemeral signers) in combination with a new @SecretExposure annotation
 
 
 ### 3.10.0 (Supreme 0.5.0) More ~~cowbell~~ targets!